requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
haystack-ai==2.12.1
pgvector-haystack==1.4.0
sentence-transformers>=3.3.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic>=2.0.0